Determine the repo path in a symlink-friendly way.
The Contributions/selflink.sh script does not symlink the .git directory,
so 'brew update' will fail because the prefix (/usr/local/) is not a git
repository.

To determine the actual location of the git repo we can get the realpath
of the Library (to expand any symlinks) and take the parent of that.