README fix ups
The point being to have less duplicate information.

I am updating the wiki and the homepage accordingly.

I am loathe to leave the "what packages are available" lines but I believe
people who are interested in such things are likely to end up at the source
root so it is appropriate.