Add `devel` to the DSL, + stable and bottle blocks
This commit adds a `devel` entry to the DSL, allowing formulae to
specify an unstable branch.

`devel` takes a block, which should contain standard `url` and `md5`
fields (and `version`, if necessary). This must come after the standard
DSL fields.

This commit also migrates over all formulae currently using `devel` to
the new syntax, as well as formulae which used `head` for non-VCS urls.

The new syntax is also available for `stable` and `bottle`. `stable` is
an option alongside the old syntax. `bottle` replaces the old syntax.

Note that the @stable ivar in Formula has been renamed to @standard,
and the @bottle ivar has been renamed to @bottle_url.

Closes #9735.

Signed-off-by: Misty De Meo <mistydemeo@gmail.com>

Conflicts:

	Library/Formula/bdw-gc.rb
	Library/Formula/boost.rb
	Library/Formula/cmake.rb
	Library/Formula/erlang.rb
	Library/Formula/fltk.rb
	Library/Formula/fox.rb
	Library/Formula/ghc.rb
	Library/Formula/guile.rb
	Library/Formula/icu4c.rb
	Library/Formula/imagemagick.rb
	Library/Formula/libmusicbrainz.rb
	Library/Formula/nginx.rb
	Library/Formula/qt.rb
	Library/Formula/scantailor.rb
	Library/Formula/wine.rb
	Library/Homebrew/formula.rb
	Library/Homebrew/formula_installer.rb